+++ unit_tests/run_linked_list.sh
@@ -2,17 +2,23 @@
 
 test_input()
 {
-	./linked_list "$2" "$3" > "$3".txt
-	diff "$3".txt txt/"$1"."$3".exp.txt > diffs/"$1"."$3".txt
-	DIFF=$(diff "$3".txt txt/"$1"."$3".exp.txt)
-	if [ "$DIFF" != "" ]
+	ERROR=$( ./linked_list "$2" "$3" 2>&1 >/dev/null )
+	if [ $? = 0 ]
 	then
-		echo -e "\033[0;31m[KO]\033[0m"
+		./linked_list "$2" "$3" > "$3".txt
+		diff "$3".txt txt/"$1"."$3".exp.txt > diffs/"$1"."$3".txt
+		DIFF=$(diff "$3".txt txt/"$1"."$3".exp.txt)
+		if [ "$DIFF" != "" ]
+		then
+			echo -e "\033[0;31m[KO]\033[0m"
+		else
+			echo -e "\033[0;32m[OK]\033[0m"
+			rm -rf diffs/"$1"."$3".txt
+		fi
+		rm -rf "$3".txt
 	else
-		echo -e "\033[0;32m[OK]\033[0m"
-		rm -rf diffs/"$1"."$3".txt
+		echo -e "\033[0;31m[ERROR]\033[0m"
 	fi
-	rm -rf "$3".txt
 }
 
 run_all_functions()
@@ -39,24 +45,41 @@ run_function()
 
 }
 
-clang testers/test_linked_list.c -L. -lasm -o linked_list
 FUNCT=(list_push_front list_size list_sort list_remove_if)
 
-if [ "$1" == "list_push_front" ]; then
-	only_funct=0
-elif [ "$1" == "list_size" ]; then
-	only_funct=1
-elif [ "$1" == "list_sort" ]; then
-	only_funct=2
-elif [ "$1" == "list_remove_if" ]; then
-	only_funct=3
-fi
 
-if [ "$1" == "" ]
-then
-	run_all_functions
-else
-	run_function "$1" $only_funct
-fi
+COMP=$( clang testers/test_linked_list.c -L. -lasm -o linked_list 2>&1 >/dev/null )
 
-rm -rf linked_list
+# if [ $? = 0 ]
+# then
+	clang testers/test_linked_list.c -L. -lasm -o linked_list
+
+	if [ "$1" == "list_push_front" ]; then
+		only_funct=0
+	elif [ "$1" == "list_size" ]; then
+		only_funct=1
+	elif [ "$1" == "list_sort" ]; then
+		only_funct=2
+	elif [ "$1" == "list_remove_if" ]; then
+		only_funct=3
+	fi
+
+	if [ "$1" == "" ]
+	then
+		run_all_functions
+	else
+		run_function "$1" $only_funct
+	fi
+
+	rm -rf linked_list
+
+# else
+# 	echo -e "\033[0;31m[DONT COMPILE]\033[0m"
+# fi
+
+
+
+#
+# the test compile only when all functions were add to the libasm.a
+# even if is a empty source file, need to fix this!
+#
